+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ### Added
 
 - `closestPierce` in browser utils
+- `getScrollbarWidth` options in browser utils
 
 ## [2.5.1]